backend/requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.1
asn1crypto==0.24.0
attrs==19.3.0
Automat==0.6.0
backcall==0.2.0
bcrypt==3.2.0
bleach==3.1.5
blinker==1.4
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cloud-init==19.4
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
cvxpy==1.0.31
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
distro-info===0.18ubuntu0.18.04.1
Django==3.1.7
django-admin-tools==0.9.1
django-cors-headers==3.7.0
django-object-actions==3.0.2
django-phonenumber-field==5.1.0
djangorestframework==3.12.4
djangorestframework-xml==2.0.0
docopt==0.6.2
ecos==2.0.7.post1
entrypoints==0.3
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.1
future==0.18.2
gevent==21.1.2
gif==3.0.0
greenlet==1.0.0
h5py==2.10.0
httplib2==0.9.2
hyperlink==17.3.1
idna==2.10
importlib-metadata==1.7.0
incremental==16.10.1
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.17.0
jsonpatch==1.16
jsonpointer==1.10
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
Keras==2.4.3
keras-utils==1.0.13
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
language-selector==0.1
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
multiprocess==0.70.9
nbconvert==5.6.1
nbformat==5.0.7
netifaces==0.10.4
nltk==3.5
notebook==6.1.1
numpy==1.19.4
oauthlib==2.0.6
opencv-python==4.2.0.34
osqp==0.6.1
packaging==20.4
PAM==0.4.2
pandas==1.1.3
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
phonenumbers==8.12.22
pickleshare==0.7.5
Pillow==8.0.1
pipreqs==0.4.10
prometheus-client==0.8.0
prompt-toolkit==3.0.6
psutil==5.8.0
psycopg2==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
pycrypto==2.6.1
Pygments==2.6.1
pygobject==3.26.1
PyJWT==1.5.3
PyMySQL==1.0.2
pyOpenSSL==17.5.0
pyparsing==2.4.7
pyperclip==1.8.0
pyrsistent==0.16.0
pyserial==3.4
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
pytictoc==1.5.0
pytz==2021.1
pyxdg==0.25
PyYAML==5.3.1
pyzmq==19.0.2
raven==6.10.0
regex==2020.5.14
requests==2.18.4
requests-unixsocket==0.1.5
scikit-learn==0.23.2
scipy==1.5.3
scs==2.1.2
seaborn==0.11.0
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.5.0
service-identity==16.0.0
Shapely==1.7.1
six==1.15.0
sklearn==0.0
SQLAlchemy==1.3.16
sqlparse==0.4.1
ssh-import-id==5.7
systemd-python==234
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
torch==1.7.0
torchvision==0.8.1
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
ttictoc==0.5.6
Twisted==17.9.0
typing-extensions==3.7.4.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.4
utils==1.0.1
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
xmltodict==0.12.0
yarg==0.1.9
zipp==3.1.0
zope.event==4.5.0
zope.interface==5.4.0
